requirements.txt
Django==3.2.9
django-crispy-forms==1.13.0
django-extensions==3.1.5
Pillow==8.4.0
psycopg2-binary==2.9.1
python-dotenv==0.19.1